week 3/week3-copilot-memory-api/Day 9/requirements.txt
fastapi
uvicorn
openai>=1.0.0
langchain
langchain-core
langchain-openai
pydantic
python-dotenv
